requirements.txt
# Core libraries
python-dotenv
rumps
sets

# Recommended GUI library
PySide6
